requirements.txt
functions-framework==3.*
nltk==3.8.1
python-dotenv==0.21.0